back/requirements.txt
falsk
flask-cors
flask-socketio